updated pos code and removed vidado project

+++ POS/POS_TEST/test.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-int main()
+unsigned short int main()
 {
 	return 0; 
 }
